исправил

+++ src/main/java/web/service/UserServiceImpl.java
@@ -3,7 +3,6 @@ package web.service;
 import org.springframework.stereotype.Service;
 import web.dao.UserDao;
 import web.model.User;
-
 import java.util.List;
 
 @Service
@@ -22,24 +21,33 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User findById(Long id) {
-        return userDao.findById(id);
+        if (userDao.findById(id) == null) {
+            return null;
+        } else {
+            return userDao.findById(id);
+        }
     }
-
 
     @Override
     public void add(User user) {
         userDao.add(user);
-
     }
 
     @Override
     public void update(User user) {
-        userDao.update(user);
-
+        if (userDao.findById(user.getId()) == null) {
+            userDao.add(user);
+        } else {
+            userDao.update(user);
+        }
     }
 
     @Override
     public void delete(Long id) {
-        userDao.delete(id);
+        if (userDao.findById(id) == null) {
+            throw new RuntimeException("User not found");
+        } else {
+            userDao.delete(id);
+        }
     }
 }
